+++ src/main/java/com/bookbla/americano/domain/member/repository/MemberPolicyRepository.java
@@ -2,7 +2,7 @@ package com.bookbla.americano.domain.member.repository;
 
 import com.bookbla.americano.domain.member.repository.entity.Member;
 import com.bookbla.americano.domain.member.repository.entity.MemberPolicy;
-import com.bookbla.americano.domain.member.repository.entity.Policy;
+import com.bookbla.americano.domain.policy.Policy;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
