+++ src/main/java/com/bookbla/americano/domain/policy/repository/PolicyRepository.java
@@ -1,6 +1,6 @@
-package com.bookbla.americano.domain.member.repository;
+package com.bookbla.americano.domain.policy.repository;
 
-import com.bookbla.americano.domain.member.repository.entity.Policy;
+import com.bookbla.americano.domain.policy.Policy;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
